Remove cpu time metrics.
