--- src/Core/CpuTimeByFeatureMetrics/DefaultActiveMetricsMeasurerStorage.cs
@@ -1,34 +1,0 @@
-// Copyright 2021 Mindbox Ltd
-// 
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// 
-//     http://www.apache.org/licenses/LICENSE-2.0
-// 
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#nullable disable
-
-using System.Threading;
-
-namespace Mindbox.DiagnosticContext.CpuTimeByFeatureMetrics;
-
-internal sealed class DefaultActiveMetricsMeasurerStorage : IActiveMetricsMeasurerStorage
-{
-	private static readonly ThreadLocal<MetricsMeasurer> _measurer = new();
-
-	public MetricsMeasurer GetActiveMeasurer()
-	{
-		return _measurer.Value;
-	}
-
-	public void SetActiveMeasurer(MetricsMeasurer newMeasurer)
-	{
-		_measurer.Value = newMeasurer;
-	}
-}
